requirements.txt
pandas
python-binance
requests 
IPython
pyinstaller
jupyter